Create TEST-tutorial-image-to-ghcr.yaml
I created a test image copying workflow (I should only need to perform this operation once, I just don't have a machine that runs docker locally). Note that the code is just a copy-paste of deploy.yaml right now (modulo a comment explaining the situation); that should be fixed in the coming commits.