Update deploy.yaml
Of course, I shouldn't be pushing two images with the same tags at the same time (namely, when I create a pull request), so I'm no longer pushing the "latest" tag here. This way, the tag with the SHA is the image of the repository build while the tag with latest is the copy of the tutorial image.